sub-commit 3 of 3; the rest of what i've been working on. probably not working quite yet, but close.

+++ kalite/securesync/management/commands/initdevice.py
@@ -1,20 +1,61 @@
 import os
+import sys
 
 from django.core.management import call_command
 from django.core.management.base import BaseCommand, CommandError
+from django.db import IntegrityError, transaction
 
 import settings
 import version
 from securesync.models import Device, DeviceMetadata, Zone, ZoneInvitation
+from securesync.views import set_as_registered
+from settings import LOG as logging
 from shared import serializers
 from utils.general import get_host_name
+
+
+@transaction.commit_on_success  # because the objects may not be in order, do in a single transaction
+def load_data_for_offline_install(in_file):
+    """
+    Receives a serialized file for import.
+    Import the file--nothing more!
+    
+    File should contain:
+    * Central server object
+    and, optionally
+    * Zone object
+    * Device and DeviceZone / ZoneInvitation objects (chain of trust)
+
+    Essentially duplicates code from securesync.device.api_client:RegistrationClient
+    """
+    assert os.path.exists(in_file), "in_file must exist."
+    with open(in_file, "r") as fp:
+        models = serializers.deserialize("versioned-json", fp.read())  # all must be in a consistent version
+    
+    # First object should be the central server.
+    try:
+        central_server = models.next().object
+    except Exception as e:
+        logging.debug("Exception loading central server object: %s" % e)
+        return
+    logging.debug("Saving object %s" % central_server)
+    assert isinstance(central_server, Device)
+    assert central_server.verify()
+    central_server.save(imported=True, is_trusted=True)
+
+    # Everything else, import as is.
+    for model in models:
+        logging.debug("Saving object %s" % model.object)
+        assert model.object.verify()
+        model.object.save(imported=True)
 
 
 class Command(BaseCommand):
     args = "\"<name of device>\" \"<description of device>\""
     help = "Initialize device with optional name and description"
 
-    install_json_file = os.path.join(settings.STATIC_ROOT, "data", "install_data.json")
+    install_json_filename = "install_data.json"
+    install_json_file = os.path.join(settings.STATIC_ROOT, "data", install_json_filename)
 
     def handle(self, *args, **options):
         if DeviceMetadata.objects.filter(is_own_device=True).count() > 0:
@@ -24,11 +65,7 @@ class Command(BaseCommand):
         description = args[1] if (len(args) >= 2 and args[1]) else ""
         data_file   = args[2] if (len(args) >= 3 and args[2]) else self.install_json_file
 
-        # Validate parameters
-        if not os.path.exists(data_file):
-            raise CommandError("Could not find resource file %s; installation will run, but may have restricted syncing functionality." % data_file)
-
-        Device.initialize_own_device(name=name, description=description)
+        own_device = Device.initialize_own_device(name=name, description=description)
         self.stdout.write("Device '%s'%s has been successfully initialized.\n"
             % (name, description and (" ('%s')" % description) or ""))
 
@@ -38,33 +75,30 @@ class Command(BaseCommand):
 
         # Now we're definitely not central server, so ... go for it!
         # Import a zone (for machines sharing zones)
-        try:
-            Zone.load_zone_for_offline_install(in_file=data_file)
-            self.stdout.write("Successfully imported zone information from %s\n" % data_file)
-            os.remove(data_file)
-        except Exception as e:
-            raise CommandError("Error importing objects: %s\n" % str(e)) 
-
-        # Generate a zone (for stand-alone machines)
-        unused_invitations = ZoneInvitation.objects.filter(device=None).exclude(private_key=None)
-        if unused_invitations:
-            invitation = unused_invitations[0]
+        if not os.path.exists(data_file):
+            sys.stderr.write("Could not find resource file %s.  This may cause warnings to appear when updating your KA Lite version." % data_file)
         else:
-            invitation = None
+            import pdb; pdb.set_trace()
+            try:
+                load_data_for_offline_install(in_file=data_file)
+                self.stdout.write("Successfully imported offline data from %s\n" % data_file)
+                if not settings.DEBUG:
+                    os.remove(data_file)
+            except Exception as e:
+                raise CommandError("Error importing offline data from %s: %s\n" % (data_file, str(e))) 
+
+        # Join a zone, either by grabbing an open invitation, or by generating one.
+        unused_invitations = ZoneInvitation.objects.filter(used_by=None).exclude(private_key=None)
+        if unused_invitations:
+            # Zone info existed in the data blob we received.  Use it to join the zone!
+            invitation = unused_invitations[0]
+            invitation.claim(used_by=own_device)
+            self.stdout.write("Successfully joined existing zone %s, using invitation %s.\n" % (invitation.zone, invitation))
+
+        else:
             # Sorry dude, you weren't invited to the party.  You'll have to have your own!
+            # Generate a zone (for stand-alone machines)
             call_command("generate_zone")
-            self.stdout.write("Successfully generated a stand-alone zone.\n") 
+            self.stdout.write("Successfully generated a stand-alone zone, and joined!.\n") 
 
-
-        # Grab the zone
-        all_zones = Zone.objects.all()
-        if len(all_zones) != 1:
-            raise CommandError("There should always be exactly one zone in the database at this time!\n")
-        zone = all_zones[0]
-
-        # "Register" locally, offline
-        try:
-            Device.get_own_device().register_offline(zone=zone, invitation=invitation)
-            self.stdout.write("Successfully registered (offline) to zone %s, using invitation '%s'\n" % (zone.name, invitation))
-        except Exception as e:
-            raise CommandError("Error completing offline registration: %s\n" % str(e))
+        set_as_registered()  # would try to sync
